+++ packages/tuya-panel-ipc-sdk/src/ty-ipc-native/cameraData.tsx
@@ -1,4 +1,3 @@
-// @ts-nocheck
 import Strings from '../ty-ipc-player/i18n';
 
 export const decodeClarityDic = {