+++ packages/tuya-panel-ipc-sdk/src/ty-ipc-native/nativeManager.tsx
@@ -1,4 +1,3 @@
-// @ts-nocheck
 /* eslint-disable consistent-return */
 /* eslint-disable no-console */
 import { TYSdk } from 'tuya-panel-kit';