+++ packages/tuya-panel-ipc-sdk/src/ty-ipc-player/components/cutScreen.tsx
@@ -1,4 +1,3 @@
-// @ts-nocheck
 import React from 'react';
 import { View, TouchableOpacity, StyleSheet, Image } from 'react-native';
 import { TYSdk, TYText, IconFont } from 'tuya-panel-kit';