+++ packages/tuya-panel-ipc-sdk/src/ty-ipc-player/index.tsx
@@ -1,4 +1,3 @@
-// @ts-nocheck
 /* eslint-disable no-underscore-dangle */
 /* eslint-disable consistent-return */
 /* eslint-disable react/sort-comp */