+++ packages/tuya-panel-ipc-sdk/src/ty-ipc-player/interFace.tsx
@@ -1,4 +1,3 @@
-// @ts-nocheck
 import Res from './res';
 import TYIpcPlayerManager from '../ty-ipc-native';
 import Strings from './i18n';