+++ packages/tuya-panel-robot-sdk/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1](https://github.com/tuya/tuya-panel-sdk/compare/@tuya-smart/tuya-panel-robot-sdk@1.0.0...@tuya-smart/tuya-panel-robot-sdk@1.0.1) (2020-12-29)
+
+**Note:** Version bump only for package @tuya-smart/tuya-panel-robot-sdk
+
+# 1.0.0 (2020-12-11)
+
+**Note:** Version bump only for package @tuya-smart/tuya-panel-robot-sdk
+
 # 1.0.0 (2020-12-02)
 
 ### Features